+++ resources/views/components/tools/filter-pills.blade.php
@@ -14,6 +14,7 @@
             </small>
 
             @foreach($this->getAppliedFiltersWithValues() as $filterSelectName => $value)
+                @php /** @var \Rappasoft\LaravelLivewireTables\Views\Filter $filter */ @endphp
                 @php($filter = $this->getFilterByKey($filterSelectName))
                 @continue(is_null($filter) || $filter->isHiddenFromPills())
                 @php( $filterPillTitle = $filter->getFilterPillTitle())
@@ -31,4 +32,3 @@
         </div>
     </div>
 @endif
-