+++ resources/views/components/tools/toolbar.blade.php
@@ -2,12 +2,12 @@
 @props([])
 @php($toolBarAttributes = $this->getToolBarAttributesBag())
 
-<div 
+<div
     {{
         $toolBarAttributes->merge()
         ->class(['md:flex md:justify-between mb-4 px-4 md:p-0' => $isTailwind && ($toolBarAttributes['default-styling'] ?? true)])
         ->class(['d-md-flex justify-content-between mb-3' => $isBootstrap && ($toolBarAttributes['default-styling'] ?? true)])
-        ->except(['default','default-styling','default-colors']) 
+        ->except(['default','default-styling','default-colors'])
     }}
 >
     <div @class([
@@ -28,7 +28,7 @@
             <x-livewire-tables::tools.toolbar.items.reorder-buttons />
         @endif
 
-        @if ($this->searchIsEnabled() && $this->searchVisibilityIsEnabled())
+        @if ($this->searchIsEnabled && $this->searchVisibilityIsEnabled)
             <x-livewire-tables::tools.toolbar.items.search-field />
         @endif
 
@@ -36,8 +36,8 @@
             <x-livewire-tables::tools.toolbar.items.filter-button />
         @endif
 
-        @if($this->hasActions && $this->showActionsInToolbar && $this->getActionsPosition == 'left')
-            <x-livewire-tables::includes.actions/>    
+        @if($this->hasActions && $this->showActionsInToolbar && $this->getActionsPosition === 'left')
+            <x-livewire-tables::includes.actions/>
         @endif
 
         @if ($this->hasConfigurableAreaFor('toolbar-left-end'))
@@ -58,8 +58,8 @@
     >
         @includeWhen($this->hasConfigurableAreaFor('toolbar-right-start'), $this->getConfigurableAreaFor('toolbar-right-start'), $this->getParametersForConfigurableArea('toolbar-right-start'))
 
-        @if($this->hasActions && $this->showActionsInToolbar && $this->getActionsPosition == 'right')
-            <x-livewire-tables::includes.actions/>    
+        @if($this->hasActions && $this->showActionsInToolbar && $this->getActionsPosition === 'right')
+            <x-livewire-tables::includes.actions/>
         @endif
 
         @if ($this->showBulkActionsDropdownAlpine() && $this->shouldAlwaysHideBulkActionsDropdownOption != true)
@@ -85,4 +85,3 @@
 )
     <x-livewire-tables::tools.toolbar.items.filter-slidedown  />
 @endif
-