+++ resources/views/includes/columns/boolean.blade.php
@@ -1,24 +1,24 @@
-@if($isToggleable && $toggleMethod != '')
-    <button wire:click="{{ $toggleMethod }}('{{ $rowPrimaryKey }}')" 
+@if($isToggleable && $toggleMethod !== '')
+    <button wire:click="{{ $toggleMethod }}('{{ $rowPrimaryKey }}')"
     @if($hasConfirmMessage) wire:confirm="{{ $confirmMessage }}" @endif
 >
 @endif
     @if ($status)
         @if ($type === 'icons')
             @if ($successValue === true)
-                <x-heroicon-o-check-circle 
+                <x-heroicon-o-check-circle
                     @class(
                         [
-                            "inline-block h-5 w-5 text-green-500" => $isTailwind,
-                            "d-inline-block text-success laravel-livewire-tables-btn-small" => $isBootstrap
+                            'inline-block h-5 w-5 text-green-500' => $isTailwind,
+                            'd-inline-block text-success laravel-livewire-tables-btn-small' => $isBootstrap
                         ]
-                    ) 
+                    )
                 />
             @else
                 <x-heroicon-o-check-circle @class(
                         [
-                            "inline-block h-5 w-5 text-red-500" => $isTailwind,
-                            "d-inline-block text-danger laravel-livewire-tables-btn-small" => $isBootstrap
+                            'inline-block h-5 w-5 text-red-500' => $isTailwind,
+                            'd-inline-block text-danger laravel-livewire-tables-btn-small' => $isBootstrap
                         ]
                     )
                 />
@@ -35,18 +35,18 @@
             @if ($successValue === false)
                 <x-heroicon-o-x-circle @class(
                         [
-                            "inline-block h-5 w-5 text-green-500" => $isTailwind,
-                            "d-inline-block text-success laravel-livewire-tables-btn-small" => $isBootstrap
+                            'inline-block h-5 w-5 text-green-500' => $isTailwind,
+                            'd-inline-block text-success laravel-livewire-tables-btn-small' => $isBootstrap
                         ]
-                    )  
+                    )
                 />
             @else
                 <x-heroicon-o-x-circle @class(
                         [
-                            "inline-block h-5 w-5 text-red-500" => $isTailwind,
-                            "d-inline-block text-danger laravel-livewire-tables-btn-small" => $isBootstrap
+                            'inline-block h-5 w-5 text-red-500' => $isTailwind,
+                            'd-inline-block text-danger laravel-livewire-tables-btn-small' => $isBootstrap
                         ]
-                    ) 
+                    )
                 />
             @endif
         @elseif ($type === 'yes-no')
@@ -57,6 +57,6 @@
             @endif
         @endif
     @endif
-@if($isToggleable && $toggleMethod != '')
+@if($isToggleable && $toggleMethod !== '')
     </button>
 @endif