+++ resources/views/includes/columns/link.blade.php
@@ -1,3 +1,5 @@
+@php /** @var \Rappasoft\LaravelLivewireTables\Views\Columns\LinkColumn $column */ @endphp
+
 <a href="{{ $path }}" {!! count($attributes) ? $column->arrayToAttributes($attributes) : '' !!}>
     @if($column->isHtml())
         {!! $title !!}