+++ src/Traits/Configuration/SearchConfiguration.php
@@ -147,20 +147,6 @@ trait SearchConfiguration
         return $this;
     }
 
-    public function setSearchPlaceholder(string $placeholder): self
-    {
-        $this->searchPlaceholder = $placeholder;
-
-        return $this;
-    }
-
-    public function setSearchFieldAttributes(array $attributes = []): self
-    {
-        $this->setCustomAttributes('searchFieldAttributes', array_merge(['default' => false, 'default-colors' => false, 'default-styling' => false], $attributes));
-
-        return $this;
-    }
-
     public function setTrimSearchString(bool $status): self
     {
         $this->trimSearchString = $status;