+++ src/Traits/Styling/HasSearchFieldStyling.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Styling;
+
+use Rappasoft\LaravelLivewireTables\Traits\Styling\Search\{HasSearchIcon, HasSearchInput};
+
+trait HasSearchFieldStyling
+{
+    use HasSearchIcon,
+        HasSearchInput;
+}